+++ test/read_it_test.dart
@@ -38,27 +38,12 @@ void main() {
     test('should provide a lazy value', () {
       final readIt = ReadIt.asNewInstance();
       int? lazyValue;
-      readIt.provideLazy<int>(() => lazyValue = 42);
+      readIt.provide<int>(() => lazyValue = 42, lazy: true);
       expect(lazyValue, isNull);
 
       final value = readIt.read<int>();
       expect(value, 42);
       expect(lazyValue, 42);
-    });
-
-    test('should provide a factory value', () {
-      final readIt = ReadIt.asNewInstance();
-      readIt.provideFactory(Counter.new);
-
-      final counter1 = readIt.read<Counter>();
-      final counter2 = readIt.read<Counter>();
-      expect(counter1, isNot(same(counter2)));
-
-      counter1.value = 43;
-      expect(counter2.value, 42);
-
-      counter2.value = 44;
-      expect(counter1.value, 43);
     });
 
     test('should be all ready', () async {